requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-WTF==1.0.1
gunicorn==20.1.0 
oauthlib==3.2.2 
python-dotenv==0.21.0
requests==2.28.1  